+++ RELEASE_NOTES.md
@@ -1,3 +1,1 @@
- * Fixed a problem with grep 3.8
- * Added ```--grep-bin``` to specify the grep binary to be used
- * Added ```--debug-headers``` to store the HTTP headers in the headers.txt file
+Checks the maximum certificate validity (397)